Organize folders into shared / features
