+++ twoframe-client/src/features/admin/hooks/useGetAttendees.ts
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 
 //this query is dependent on the first query which grabs the tournament
 //once the tournament is succesfully grabbed, we can perform a fetch of the attendees
-export function useAttendees(code: string, tournamentId?: string) {
+export function useGetAttendees(code: string, tournamentId?: string) {
   return useQuery({
     queryKey: ["attendees", code],
     queryFn: async () => {
